New feature
Pickr now can be shown inline. When pickr is attached to a DOM element that's not an input, it will activate the inline mode automatically.

It can also be forced with the option `inline` to `true`.

To be able to catch the value when on inline mode, pickr now triggers a `pickr:selected` event with the value as a parameter.

Note that the event is triggered both on inline and in normal mode.

#Fixes 8